+++ app/javascript/components/widgets/climate/emissions/selectors.js
@@ -137,6 +137,6 @@ export const parseSentence = createSelector(
 
 export default createStructuredSelector({
   data: parseData,
-  dataConfig: parseConfig,
+  config: parseConfig,
   sentence: parseSentence
 });
